phase1_api_basics/requirements.txt
# ============================================
# Phase 1: API 基础 - 依赖包
# ============================================
# 安装: pip install -r requirements.txt

# ============================================
# LLM 客户端
# ============================================
openai==1.10.0
anthropic==0.18.1
zhipuai==2.1.0.20240119
dashscope==1.14.1

# ============================================
# HTTP & 异步
# ============================================
requests==2.31.0
aiohttp==3.9.1

# ============================================
# 工具库
# ============================================
python-dotenv==1.0.0
tiktoken==0.5.2  # Token 计算

# ============================================
# 日志
# ============================================
loguru==0.7.2

# ============================================
# 测试
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
